refactor(lucene): 优化内存计算和线程数设置
- 在 FileProperties 中新增 determineEffectiveNGramLength 方法，用于计算有效的 ngram 最大内容长度
- 在 LuceneService 中使用 Math.toIntExact 代替直接类型转换，提高代码安全性

+++ src/main/java/com/jmal/clouddisk/lucene/LuceneService.java
@@ -130,7 +130,7 @@ public class LuceneService {
             // 获取可用处理器数量
             int smallProcessors = Runtime.getRuntime().availableProcessors() - 3;
             // 设置线程数, 假设每个线程占用内存为500M
-            int maxSmallProcessors = (int) ((maxMemory / BYTES_PER_MB) / MEMORY_PER_SMALL_THREAD_MB);
+            int maxSmallProcessors =  Math.toIntExact((maxMemory / BYTES_PER_MB) / MEMORY_PER_SMALL_THREAD_MB);
             if (smallProcessors > maxSmallProcessors) {
                 smallProcessors = maxSmallProcessors;
             }
